requirements.txt
# Core Dependencies
openai>=1.0.0
sounddevice>=0.4.6
scipy>=1.10.0
elevenlabs>=0.2.26
webrtcvad-wheels>=2.0.10
pyaudio>=0.2.13

# Additional Utilities
numpy>=1.25.2
requests>=2.31.0
python-dotenv>=1.0.0
keyboard>=0.13.5
pyttsx3>=2.90
python-dateutil>=2.8.2
